# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: User categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category removed
        "404":
          description: Not found
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      responses:
        "200":
          description: User credit cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Created credit card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: Credit card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
    patch:
      operationId: updateCard
      tags: [cards]
      parameters:
        - $ref: "#/components/parameters/CardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: Updated credit card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
    delete:
      operationId: deleteCard
      tags: [cards]
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "204":
          description: Credit card removed
        "404":
          description: Not found
  /cards/{id}/pay-invoice:
    post:
      operationId: payCardInvoice
      tags: [cards]
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: Credit card invoice paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found
  /cards/{id}/history:
    get:
      operationId: getCardHistory
      tags: [cards]
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: Current card invoice history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardHistoryItem"
        "404":
          description: Not found
...
  /account:
    delete:
      operationId: deleteAccount
      tags: [account]
      summary: Permanently delete the authenticated account
      responses:
        "204":
          description: Account and all financial data removed
        "401":
          description: Unauthorized
  /account/profile:
    patch:
      operationId: updateAccountProfile
      tags: [account]
      summary: Update the authenticated user's profile name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "204":
          description: Profile updated
        "400":
          description: Invalid profile name
        "401":
          description: Unauthorized
  /account/profile-image:
    patch:
      operationId: updateAccountProfileImage
      tags: [account]
      summary: Update the authenticated user's profile image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileImageUpdate"
      responses:
        "204":
          description: Profile image updated
        "400":
          description: Invalid image
        "401":
          description: Unauthorized
components:
  parameters:
...
    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CardId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
      required:
        - firstName
    ProfileImageUpdate:
      type: object
      properties:
        data:
          type: string
          minLength: 1
        mimeType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
      required:
        - data
        - mimeType
...
    Category:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        createdAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }
...
    InvoiceStatus:
      type: string
      enum: [open, closed]
    Card:
      type: object
      required: [id, name, dueDay, closingDay, currentInvoiceAmount, availableLimit, invoiceStatus, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        dueDay: { type: integer, minimum: 1, maximum: 31 }
        closingDay: { type: integer, minimum: 1, maximum: 31 }
        currentInvoiceAmount: { type: number, minimum: 0 }
        availableLimit: { type: number, minimum: 0, nullable: true }
        invoiceStatus:
          $ref: "#/components/schemas/InvoiceStatus"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CardInput:
      type: object
      required: [name, dueDay, closingDay]
      properties:
        name: { type: string, minLength: 1 }
        dueDay: { type: integer, minimum: 1, maximum: 31 }
        closingDay: { type: integer, minimum: 1, maximum: 31 }
        availableLimit: { type: number, minimum: 0, nullable: true }
    CardUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        dueDay: { type: integer, minimum: 1, maximum: 31 }
        closingDay: { type: integer, minimum: 1, maximum: 31 }
        availableLimit: { type: number, minimum: 0, nullable: true }
    CardHistoryItem:
      type: object
      required: [id, kind, description, amount, date, paymentStatus]
      properties:
        id: { type: string }
        kind: { type: string, enum: [transaction, closure] }
        description: { type: string }
        amount: { type: number, minimum: 0 }
        date: { type: string, format: date }
        paymentStatus: { type: string, enum: [paid, unpaid] }
        categoryName: { type: string, nullable: true }